# docs/frontleaves_plugin_swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/servers": {
            "get": {
                "description": "分页查询服务器列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "超管-服务器管理接口"
                ],
                "summary": "[超管] 查询服务器列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apiServer.ServerListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的服务器记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "超管-服务器管理接口"
                ],
                "summary": "[超管] 创建服务器",
                "parameters": [
                    {
                        "description": "创建服务器请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apiServer.CreateServerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apiServer.ServerResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        },
        "/admin/servers/{id}": {
            "get": {
                "description": "根据服务器 ID 查询详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "超管-服务器管理接口"
                ],
                "summary": "[超管] 查询服务器详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务器ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apiServer.ServerResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "服务器不存在",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "更新指定服务器的信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "超管-服务器管理接口"
                ],
                "summary": "[超管] 更新服务器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务器ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新服务器请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apiServer.UpdateServerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apiServer.ServerResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "服务器不存在",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "硬删除指定的服务器记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "超管-服务器管理接口"
                ],
                "summary": "[超管] 删除服务器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务器ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "服务器不存在",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        },
        "/admin/servers/{id}/enabled": {
            "put": {
                "description": "启用或禁用服务器",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "超管-服务器管理接口"
                ],
                "summary": "[超管] 设置服务器启用状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务器ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "设置启用状态请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apiServer.SetServerEnabledRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apiServer.ServerResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "服务器不存在",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        },
        "/admin/servers/{id}/public": {
            "put": {
                "description": "设置服务器是否对外公开可见",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "超管-服务器管理接口"
                ],
                "summary": "[超管] 设置服务器公开状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务器ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "设置公开状态请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apiServer.SetServerPublicRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apiServer.ServerResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "服务器不存在",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apiServer.CreateServerRequest": {
            "type": "object",
            "required": [
                "display_name",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
        "apiServer.ServerListResponse": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apiServer.ServerResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "apiServer.ServerResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_enabled": {
                    "type": "boolean"
                },
                "is_public": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "apiServer.SetServerEnabledRequest": {
            "type": "object",
            "properties": {
                "is_enabled": {
                    "type": "boolean"
                }
            }
        },
        "apiServer.SetServerPublicRequest": {
            "type": "object",
            "properties": {
                "is_public": {
                    "type": "boolean"
                }
            }
        },
        "apiServer.UpdateServerRequest": {
            "type": "object",
            "required": [
                "display_name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
        "xBase.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "context": {
                    "type": "string"
                },
                "data": {},
                "error_message": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                },
                "overhead": {
                    "type": "integer"
                }
            }
        }
    }
}